fix: Fix pagination tests
