--- packages/overdrive/lib/components/Pagination/Loading.tsx
@@ -1,57 +1,0 @@
-import { ChevronLeftIcon, ChevronRightIcon } from '@autoguru/icons';
-import clsx from 'clsx';
-import * as React from 'react';
-import { FunctionComponent } from 'react';
-import { useStyles } from 'react-treat';
-
-import { Box } from '../Box';
-import { Icon } from '../Icon';
-import { Bubble } from './Bubble';
-import * as styleRefs from './Pagination.treat';
-
-export interface Props {
-	className?: string;
-	placeholderBubblesNum?: number;
-}
-
-export const LoadingComponent: FunctionComponent<Props> = ({
-	placeholderBubblesNum = 3,
-}) => {
-	const styles = useStyles(styleRefs);
-	const disabledChevCls = clsx(
-		styles.activeItem.default,
-		styles.chevron.default,
-		styles.chevron.disabled,
-	);
-	return (
-		<Box
-			is="span"
-			display="flex"
-			alignItems="center"
-			flexDirection="row"
-			flexWrap="wrap"
-			justifyContent="flexStart">
-			<span className={disabledChevCls}>
-				<Icon
-					className={styles.chevron.icon}
-					size="medium"
-					icon={ChevronLeftIcon}
-				/>
-			</span>
-			{new Array(placeholderBubblesNum).fill('').map((_, index) => (
-				<Bubble
-					key={index}
-					children=""
-					className={styles.bubble.disabled}
-				/>
-			))}
-			<span className={disabledChevCls}>
-				<Icon
-					className={styles.chevron.icon}
-					size="medium"
-					icon={ChevronRightIcon}
-				/>
-			</span>
-		</Box>
-	);
-};
